+++ modules/yahooAdsBidAdapter.md
